Development (#37)
* Fix merge conflicts

This commit fixes a bunch of compounded merge conflicts

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Recode everything and split server/client

This commit re-organizes the code, splits the main program
into a server and client since we don't want to have to multi-thread,
and adds a setup script for placing each nic in its own network namespace.

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add working server and client code

This commit fixes some issues and cleans up the server
and client. This is a buggy yet working solution.

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add routes to setup script and clean up code

This commit adds important routes to the setup steps
so that the RDMA driver knows the proper routes for
communicating

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Factory design pattern example

This commit adds the factory design pattern example to my research
repository

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add documentation for factory method

This commit adds more documentation and a README for the factory method pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Clean RDMAConnection

This commit updates header guards and syntax for the RDMA connection class

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add abstract factory outline

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add appliance and furniture classes

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add README for abstract factory method

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add builder skeleton code and refactor repo

This commit gives an outline for the builder design pattern
along with refactoring the repo's design patterns into their
descriptive directories

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Finish Builder and Add Main

This commit finishes the Builder example design pattern
along with adding a main driver

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add README for Builder

This commit adds a README for the Builder pattern code example

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Prototype Pattern

This commit adds the Prototype Pattern example. It also fixes
issues in the other design pattern examples

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Singleton example

This commit adds the Singleton design pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add README for Singleton

This commit adds the README for the singleton pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add outline for adapter and clean up

This commit adds the outline for the adapter design pattern
along with cleaning existing patterns

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Adapter implementation

This commit adds the implementation of the Adapter
pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Update READMEs

This commit updates READMEs for prototype, singleton, and adapter
patterns

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Bridge pattern

This commit adds the Bridge design pattern code example

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Bridge README

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Composite Pattern Code

This commit implements the Composite Design Pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Update Composite pattern code

This commit updates the Composite pattern example to conform it
to the pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Decorator Pattern Example

This commit adds the example for the Decorator design pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add README for Decorator

This commit adds the README for the Decorator design pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Facade Skeleton

This commit adds a skeleton for the Facade pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Finish Facade Design Pattern Example

This commit finishes the example for the Facade pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Add Facade pattern README

This commit adds the Facade Design Pattern's README

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

* Finish Proxy pattern and add README

THis commit finishes the Proxy design pattern
along with adding documentation

Signed-off-by: leonmatt <matthew.leon@intel.com>

* Fix bug in request.h

This commit adds a missing nullptr check for one of the RequestHandlers

Signed-off-by: leonmatt <matthew.leon@intel.com>

* Add Command pattern

This commit adds the example code for the command design pattern

Signed-off-by: Matthew Leon <matthew.leon@intel.com>

---------

Signed-off-by: Matthew Leon <matthew.leon@intel.com>
Signed-off-by: leonmatt <matthew.leon@intel.com>